+++ CHANGELOG.md
@@ -16,6 +16,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 * Add `disk_usage` argument to kwcoco stats.
 
+### Fixed
+* Fix detection evaluation crash when inputs are empty
+
 ## Version 0.8.9 - Released 2025-11-25
 
 ### Added